Move Snapshot related classes to its own package
