+++ src/main/scala/akka/persistence/jdbc/dao/bytea/snapshot/ByteArraySnapshotSerializer.scala
@@ -1,7 +1,7 @@
-package akka.persistence.jdbc.dao.bytea
+package akka.persistence.jdbc.dao.bytea.snapshot
 
 import akka.persistence.SnapshotMetadata
-import akka.persistence.jdbc.dao.bytea.SnapshotTables.SnapshotRow
+import akka.persistence.jdbc.dao.bytea.snapshot.SnapshotTables.SnapshotRow
 import akka.persistence.jdbc.serialization.SnapshotSerializer
 import akka.persistence.serialization.Snapshot
 import akka.serialization.Serialization
