+++ src/main/scala/akka/persistence/jdbc/dao/bytea/snapshot/SnapshotTables.scala
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package akka.persistence.jdbc.dao.bytea
+package akka.persistence.jdbc.dao.bytea.snapshot
 
-import akka.persistence.jdbc.dao.bytea.SnapshotTables._
 import akka.persistence.jdbc.config.SnapshotTableConfiguration
+import akka.persistence.jdbc.dao.bytea.snapshot.SnapshotTables._
 
 object SnapshotTables {
   case class SnapshotRow(persistenceId: String, sequenceNumber: Long, created: Long, snapshot: Array[Byte])
