+++ src/test/scala/akka/persistence/jdbc/dao/bytea/snapshot/SnapshotTablesTest.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package akka.persistence.jdbc.dao.bytea
+package akka.persistence.jdbc.dao.bytea.snapshot
 
 import akka.persistence.jdbc.dao.TablesTestSpec
 import slick.driver.JdbcProfile
